% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_variable.R
\name{add_variable_monthweek}
\alias{add_variable_monthweek}
\title{Add time variables to a dataframe based on existing datetime variable}
\usage{
add_variable_monthweek(df, datevar = "date", ...)
}
\arguments{
\item{df}{A dataframe.}

\item{datevar}{A char string of the date variable.}

\item{\dots}{Arguments ... to be implemented in the future}
}
\value{
\item{df}{Returns a df containing the created variables: month, month.abb, monthday, week, weekday, etc.}
}
\description{
Add time variables to a dataframe based on existing datetime variable
}
