% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_list.R
\name{append_list}
\alias{append_list}
\title{Append R object to a list.}
\usage{
append_list(l, obj, obj.name, overwrite = TRUE)
}
\arguments{
\item{l}{A R-list.}

\item{obj}{R object to be appended to the list.}

\item{obj.name}{R object name to be used to the list.}

\item{overwrite}{logical value indicates whether overwrite object with the same name.}
}
\value{
\item{list_new}
}
\description{
Append R object to a list.
}
\examples{
l <- list(c("a", "b"))
x <- list(c(1:5))
append_list(l, x, "X")
}
\seealso{
\code{\link{hcl}}
}
