% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mtest.R
\name{cor_mtest}
\alias{cor_mtest}
\title{Significance test for each pair of input variables}
\usage{
cor_mtest(mat, ...)
}
\arguments{
\item{mat}{Maxtrix input with each column as a variable.}

\item{\dots}{Arguments passed to \code{\link[stats]{cor.test}}.}
}
\value{
\item{p.mat}{Returns a matrix containing the p-values of correlations}
}
\description{
Significance test which produces p-values and confidence intervals for each pair of input variables.
}
\seealso{
Function \code{\link{cor.test}}
}
