% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{explorer}
\alias{explorer}
\title{Open a file or folder}
\usage{
explorer(path)
}
\arguments{
\item{path}{The path string to a file or directory.}
}
\description{
\code{explorer} opens a file or folder on your Windows, Mac or Linux computer.
}
\examples{
explorer(getwd())
explorer("C:/Users/")
explorer("/Users/")
}
\seealso{
\code{\link[base]{shell}} and \code{\link[base]{system}}
}
