% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_list.R
\name{append_list}
\alias{append_list}
\title{Append r object to a list}
\usage{
append_list(list, obj, obj.name, overwrite = TRUE)
}
\arguments{
\item{list}{R list.}

\item{obj}{R object to be appended to the list.}

\item{obj.name}{R object name to be used to the list.}
}
\value{
\item{list_new}
}
\description{
Append r object to a list
}
\seealso{
\code{\link{hcl}}
}
