% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{explorer}
\alias{explorer}
\title{Open a file or folder on your computer, supports Windows, Mac and Linux.}
\usage{
explorer(path)
}
\arguments{
\item{path}{The path string to a file or directory.}
}
\description{
Open a file or folder on your computer, supports Windows, Mac and Linux.
}
\examples{
explorer("C:/Users/")
explorer("/Users/kang")
}
