% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Get OS imformation.}
\usage{
get_os()
}
\value{
\item{os}{Returns a string indicating the operating systems, 'windows', 'linux', or 'osx'.}
}
\description{
Get OS imformation.
}
\examples{
get_os()
}
\seealso{
Function \code{\link{which_os}}
}
