% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_colour.R
\name{ggplotColours}
\alias{ggplotColours}
\title{Use the ggplot style colors}
\usage{
ggplotColours(n = 1, h = c(0, 360) + 15)
}
\arguments{
\item{n}{An integer to define the number of colors to be generated .}
}
\description{
Create a vector of colors of the ggplot style.
}
\examples{
ggplotColours(3)
}
\seealso{
\code{\link{hcl}}
}
